fanclubBackend/requirements.txt
aioredis==1.3.1
asgiref==3.3.1
async-timeout==3.0.1
attrs==20.3.0
autobahn==20.12.3
Automat==20.2.0
cachetools==4.2.0
certifi==2020.12.5
cffi==1.14.4
channels==3.0.3
channels-redis==3.2.0
chardet==4.0.0
constantly==15.1.0
cryptography==3.3.1
daphne==3.0.1
Django==3.1.5
django-rest-framework==0.1.0
djangorestframework==3.12.2
google-api-core==1.25.0
googleapis-common-protos==1.52.0
hiredis==1.1.0
httplib2==0.18.1
hyperlink==21.0.0
idna==2.10
incremental==17.5.0
msgpack==1.0.2
oauthlib==3.1.0
Pillow==8.1.0
protobuf==3.14.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
PyHamcrest==2.0.2
pyOpenSSL==20.0.1
pytz==2020.5
requests==2.25.1
requests-oauthlib==1.3.0
rsa==4.7
service-identity==18.1.0
six==1.15.0
sqlparse==0.4.1
Twisted==20.3.0
txaio==20.12.1
uritemplate==3.0.1
urllib3==1.26.2
zope.interface==5.2.0
